# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.UpsertAddressRequest:
    properties:
      full_address:
        maxLength: 500
        minLength: 5
        type: string
      is_primary:
        type: boolean
      phone:
        maxLength: 20
        minLength: 5
        type: string
      recipient_name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - full_address
    - phone
    - recipient_name
    type: object
...
  response.ErrorCode:
    enum:
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - CONFLICT
    - UNPROCESSABLE_ENTITY
    - INTERNAL_SERVER_ERROR
    - VALIDATION_ERROR
    - TOKEN_EXPIRED
    - TOKEN_INVALID
    type: string
    x-enum-varnames:
    - ErrCodeBadRequest
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeNotFound
    - ErrCodeConflict
    - ErrCodeUnprocessable
    - ErrCodeInternalServerError
    - ErrCodeValidation
    - ErrCodeTokenExpired
    - ErrCodeTokenInvalid
  response.Pagination:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      error_code:
        $ref: '#/definitions/response.ErrorCode'
      list_data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.Pagination'
      status:
        type: boolean
    type: object
info:
  contact:
    name: SEAPEDIA Dev
  description: |-
    Backend API marketplace multi-role (Admin, Seller, Buyer, Driver) — COMPFEST 18.
    Versi yang tampil di UI diambil dari config.Version (di-set saat runtime).
    Login pakai email. Otorisasi mengikuti ACTIVE ROLE pada JWT (pilih via /auth/select-role).
  termsOfService: http://swagger.io/terms/
  title: SEAPEDIA API
  version: "1.0"
paths:
...
  /admin/orders:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Monitoring order (admin)
      tags:
      - Admin
  /admin/overdue-orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Order yang sedang overdue (admin)
      tags:
      - Admin
  /admin/overdue/run:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Jalankan overdue handling manual (admin)
      tags:
      - Admin
  /admin/products:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Monitoring produk (admin)
      tags:
      - Admin
...
  /admin/simulate/advance-day:
    post:
      consumes:
      - application/json
      parameters:
      - description: Jumlah hari
        in: body
        name: body
        required: true
        schema:
          properties:
            days:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Majukan waktu N hari + jalankan overdue (admin)
      tags:
      - Admin
  /admin/simulate/now:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Lihat waktu virtual saat ini (admin)
      tags:
      - Admin
  /admin/stores:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Monitoring toko (admin)
      tags:
      - Admin
  /admin/summary:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Ringkasan monitoring marketplace (admin)
      tags:
      - Admin
  /admin/users:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Monitoring users (admin)
      tags:
      - Admin
...
  /buyer/addresses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Daftar alamat (buyer)
      tags:
      - Buyer
    post:
      consumes:
      - application/json
      parameters:
      - description: Alamat
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertAddressRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tambah alamat (buyer)
      tags:
      - Buyer
  /buyer/addresses/{id}:
    delete:
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus alamat (buyer)
      tags:
      - Buyer
    put:
      consumes:
      - application/json
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      - description: Alamat
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Ubah alamat (buyer)
      tags:
      - Buyer
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"